impl weatherRepository

+++ app/src/main/java/com/example/exoweather/feature/weather/domain/model/Weather.kt
@@ -2,6 +2,6 @@ package com.example.exoweather.feature.weather.domain.model
 
 data class Weather(
     val cityName: String,
-    val temperature: Int,
+    val temperature: Double,
     val weatherDescription: String
 )
